changed crdfilter
