+++ go.mod
@@ -9,7 +9,7 @@ replace (
 
 require (
 	github.com/layer5io/meshery-adapter-library v0.1.23
-	github.com/layer5io/meshkit v0.2.28
+	github.com/layer5io/meshkit v0.2.29
 	github.com/layer5io/service-mesh-performance v0.3.2
 	helm.sh/helm/v3 v3.6.3 // indirect
 	k8s.io/apimachinery v0.21.0
